// src/proto.d.ts
export interface IMessage {
    /** Message userId */
    userId?: string | null;

    /** Message messageId */
    messageId?: number | Long | null;

    /** Message last */
    last?: number | null;

    /** Message token */
    token?: string | null;

    /** Message userChannel */
    userChannel?: string | null;

    /** Message vpsToken */
    vpsToken?: string | null;

    /** Устарело с версии 3. */
    devContext?: IDevContext[] | null;

    /** Message messageName */
    messageName?: string | null;

    /** Message version */
    version?: number | null;

    /** Message voice */
    voice?: IVoice | null;

    /** Message text */
    text?: IText | null;

    /** Message systemMessage */
    systemMessage?: ISystemMessage | null;

    /** Message legacyDevice */
    legacyDevice?: ILegacyDevice | null;

    /** Message settings */
    settings?: ISettings | null;

    /** Message status */
    status?: IStatus | null;

    /** Message device */
    device?: IDevice | null;

    /** Message bytes */
    bytes?: IBytes | null;

    /** Message initialSettings */
    initialSettings?: IInitialSettings | null;
}

/** Represents a Message. */
export class Message implements IMessage {
    /**
     * Constructs a new Message.
     * @param [properties] Properties to set
     */
    constructor(properties?: IMessage);

    /** Message userId. */
    public userId: string;

    /** Message messageId. */
    public messageId: number | Long;

    /** Message last. */
    public last: number;

    /** Message token. */
    public token: string;

    /** Message userChannel. */
    public userChannel: string;

    /** Message vpsToken. */
    public vpsToken: string;

    /** Устарело с версии 3. */
    public devContext: IDevContext[];

    /** Message messageName. */
    public messageName: string;

    /** Message version. */
    public version: number;

    /** Message voice. */
    public voice?: IVoice | null;

    /** Message text. */
    public text?: IText | null;

    /** Message systemMessage. */
    public systemMessage?: ISystemMessage | null;

    /** Message legacyDevice. */
    public legacyDevice?: ILegacyDevice | null;

    /** Message settings. */
    public settings?: ISettings | null;

    /** Message status. */
    public status?: IStatus | null;

    /** Message device. */
    public device?: IDevice | null;

    /** Message bytes. */
    public bytes?: IBytes | null;

    /** Message initialSettings. */
    public initialSettings?: IInitialSettings | null;

    /** Message content. */
    public content?:
        | 'voice'
        | 'text'
        | 'systemMessage'
        | 'legacyDevice'
        | 'settings'
        | 'status'
        | 'device'
        | 'bytes'
        | 'initialSettings';

    /**
     * Creates a new Message instance using the specified properties.
     * @param [properties] Properties to set
     * @returns Message instance
     */
    public static create(properties?: IMessage): Message;

    /**
     * Encodes the specified Message message. Does not implicitly {@link Message.verify|verify} messages.
     * @param message Message message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IMessage, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified Message message, length delimited. Does not implicitly {@link Message.verify|verify} messages.
     * @param message Message message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: IMessage, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a Message message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns Message
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: $protobuf.Reader | Uint8Array, length?: number): Message;

    /**
     * Decodes a Message message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns Message
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: $protobuf.Reader | Uint8Array): Message;

    /**
     * Verifies a Message message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): string | null;

    /**
     * Creates a Message message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns Message
     */
    public static fromObject(object: { [k: string]: any }): Message;

    /**
     * Creates a plain object from a Message message. Also converts values to other types if specified.
     * @param message Message
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: Message, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this Message to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

/** Properties of an InitialSettings. */
export interface IInitialSettings {
    /** InitialSettings userId */
    userId?: string | null;

    /** InitialSettings userChannel */
    userChannel?: string | null;

    /** InitialSettings device */
    device?: IDevice | null;

    /** InitialSettings settings */
    settings?: ISettings | null;

    /** InitialSettings locale */
    locale?: string | null;
}
// ...
export interface IDevice {
    /** Device platformType */
    platformType?: string | null;

    /** Device platformVersion */
    platformVersion?: string | null;

    /** Обязательно. Пример, SBERBOX */
    surface?: string | null;

    /** Device surfaceVersion */
    surfaceVersion?: string | null;

    /** Device features */
    features?: string | null;

    /** Device capabilities */
    capabilities?: string | null;

    /** Device deviceId */
    deviceId?: string | null;

    /** Device deviceManufacturer */
    deviceManufacturer?: string | null;

    /** Device deviceModel */
    deviceModel?: string | null;

    /** Device additionalInfo */
    additionalInfo?: string | null;
}
// ...
export interface ISettings {
    /** Settings dubbing */
    dubbing?: number | null;

    /** Settings echo */
    echo?: number | null;

    /** Settings ttsEngine */
    ttsEngine?: string | null;

    /** Settings asrEngine */
    asrEngine?: string | null;

    /** Settings asrAutoStop */
    asrAutoStop?: number | null;

    /** Settings devMode */
    devMode?: number | null;

    /** Settings authConnector */
    authConnector?: string | null;

    /** Settings surface */
    surface?: string | null;
}
// ...
export interface ILegacyDevice {
    /** LegacyDevice clientType */
    clientType?: string | null;

    /** LegacyDevice channel */
    channel?: string | null;

    /** LegacyDevice channelVersion */
    channelVersion?: string | null;

    /** LegacyDevice platformName */
    platformName?: string | null;

    /** LegacyDevice platformVersion */
    platformVersion?: string | null;

    /** LegacyDevice sdkVersion */
    sdkVersion?: string | null;

    /** LegacyDevice protocolVersion */
    protocolVersion?: string | null;
}
// ...
export interface IVoice {
    /** Voice data */
    data?: Uint8Array | null;
}
// ...
export interface IText {
    /** Text data */
    data?: string | null;

    /** Text type */
    type?: string | null;
}
// ...
export interface ISystemMessage {
    /** SystemMessage data */
    data?: string | null;
}

/** Represents a SystemMessage. */
export class SystemMessage implements ISystemMessage {
    /**
     * Constructs a new SystemMessage.
     * @param [properties] Properties to set
     */
    constructor(properties?: ISystemMessage);

    /** SystemMessage data. */
    public data: string;

    /**
     * Creates a new SystemMessage instance using the specified properties.
     * @param [properties] Properties to set
     * @returns SystemMessage instance
     */
    public static create(properties?: ISystemMessage): SystemMessage;

    /**
     * Encodes the specified SystemMessage message. Does not implicitly {@link SystemMessage.verify|verify} messages.
     * @param message SystemMessage message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: ISystemMessage, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified SystemMessage message, length delimited. Does not implicitly {@link SystemMessage.verify|verify} messages.
     * @param message SystemMessage message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: ISystemMessage, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a SystemMessage message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns SystemMessage
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: $protobuf.Reader | Uint8Array, length?: number): SystemMessage;

    /**
     * Decodes a SystemMessage message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns SystemMessage
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: $protobuf.Reader | Uint8Array): SystemMessage;

    /**
     * Verifies a SystemMessage message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): string | null;

    /**
     * Creates a SystemMessage message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns SystemMessage
     */
    public static fromObject(object: { [k: string]: any }): SystemMessage;

    /**
     * Creates a plain object from a SystemMessage message. Also converts values to other types if specified.
     * @param message SystemMessage
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: SystemMessage, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this SystemMessage to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

/** Properties of a Status. */
export interface IStatus {
    /** Status code */
    code?: number | null;

    /** Status description */
    description?: string | null;

    /** Status technicalDescription */
    technicalDescription?: string | null;
}
// ...
export interface IBytes {
    /** Bytes data */
    data?: Uint8Array | null;

    /** Bytes desc */
    desc?: string | null;
}
// ...
export interface IDevContext {
    /** DevContext name */
    name?: string | null;

    /** DevContext timestampMs */
    timestampMs?: number | Long | null;

    /** DevContext data */
    data?: string | null;
}